+++ src/components/shared/Header/NavMenu.component.jsx
@@ -1,13 +1,15 @@
-import { memo } from "react";
+import { memo, useMemo } from "react";
+import { useInventoryContext } from "../../../context/InventoryContext";
 import useToggle from "../../../hooks/useToggle";
 import FieldSelector from "../../cards/FieldSelector.component";
 
-function NavMenu({
-  visibleFields,
-  onToggleField,
-  onResetFields,
-  archivedItemsExist,
-}) {
+function NavMenu() {
+    const inventory = useInventoryContext();
+  const { items, visibleFields, toggleField, resetFields } = inventory;
+  const archivedItemsExist = useMemo(
+    () => items.some((item) => item.Status === "archived"),
+    [items],
+  );
   const [showFieldSelector, toggleFieldSelector, , closeFieldSelector] =
     useToggle(false);
 
@@ -48,8 +50,8 @@ function NavMenu({
       {showFieldSelector && (
         <FieldSelector
           visibleFields={visibleFields}
-          onToggleField={onToggleField}
-          onResetFields={onResetFields}
+          onToggleField={toggleField}
+          onResetFields={resetFields}
           onClose={closeFieldSelector}
         />
       )}