+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Refactor Header component to use context for inventory data and simplify props
+- Refactor NavMenu component to use context for inventory data and simplify props
 - Refactor MainContainer to use context for inventory management and update tests accordingly
 
 ### Deprecated